Make `Provers` and related methods public
